Splits trackers to Shoebox
